requirements.txt
Django
django-sslserver
python-decouple
django-import-export
django-celery-beat
django-celery-results
celery
eventlet